Replace ints with doubles

+++ ConsoleCalc/ConsoleCalc/Calc.cs
@@ -8,22 +8,22 @@ namespace ConsoleCalc
 {
     public class Calc
     {
-        public int Add(int x, int y)
+        public double Add(double x, double y)
         {
             return x + y;
         }
 
-        public int Sub(int x, int y)
+        public double Sub(double x, double y)
         {
             return x - y;
         }
 
-        public int Mul(int x, int y)
+        public double Mul(double x, double y)
         {
             return x * y;
         }
 
-        public int Div(int x, int y)
+        public double Div(double x, double y)
         {
             if (y != 0)
                 return x / y;
@@ -32,12 +32,12 @@ namespace ConsoleCalc
                 return 0;
         }
 
-        public int Max(int x, int y)
+        public double Max(double x, double y)
         {
             return Math.Max(x, y);
         }
 
-        public int Min(int x, int y)
+        public double Min(double x, double y)
         {
             return Math.Min(x, y);
         }
